[ft_printf.c] has more than 25 lines, but paco is good; can't say the same about paco --s

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: nbidal <nbidal@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/27 15:08:17 by nbidal            #+#    #+#             */
-/*   Updated: 2024/02/28 16:42:21 by nbidal           ###   ########.fr       */
+/*   Updated: 2024/02/28 17:05:54 by nbidal           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 int	ft_printf(const char *format, ...);
 int	ft_putchar(int c);
 int	ft_puthex_up(unsigned int n);
-int	ft_puthex(unsigned int n);
+int	ft_puthex(unsigned long int n);
 int	ft_putnbr_u(unsigned int n);
 int	ft_putnbr(int n);
 int	ft_putptr(void *p);
